refactor(fctx): merge keys.go into middleware.go
