--- fctx/keys.go
@@ -1,41 +1,0 @@
-// Copyright © 2026 Ingka Holding B.V. All Rights Reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// You may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package fctx
-
-import (
-	"context"
-
-	"go.opentelemetry.io/otel/trace"
-	"go.uber.org/zap"
-)
-
-type loggerKey struct{}
-type tracerKey struct{}
-type requestIDKey struct{}
-
-// WithLogger stores the logger in ctx.
-func WithLogger(ctx context.Context, l *zap.Logger) context.Context {
-	return context.WithValue(ctx, loggerKey{}, l)
-}
-
-// WithTracer stores the tracer in ctx.
-func WithTracer(ctx context.Context, t trace.Tracer) context.Context {
-	return context.WithValue(ctx, tracerKey{}, t)
-}
-
-// WithRequestID stores the request ID in ctx.
-func WithRequestID(ctx context.Context, id string) context.Context {
-	return context.WithValue(ctx, requestIDKey{}, id)
-}
